pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-httpauth>=4.8.0",
    "flask>=3.1.0",
    "flask-login>=0.6.3",
    "flask-restful>=0.3.10",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=1.69.0",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.40",
    "werkzeug>=3.1.3",
    "pyjwt>=2.10.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "spacy>=3.8.4",
    "flask-wtf>=1.2.2",
    "wtforms>=3.2.1",
    "numpy>=2.2.4",
    "scikit-learn>=1.6.1",
    "joblib>=1.4.2",
    "sqlalchemy-utils>=0.41.2",
    "pandas>=2.2.3",
    "tensorflow>=2.14.0",
    "prefect>=3.2.15",
    "dask>=2025.3.0",
    "matplotlib>=3.10.1",
]
